Fix properties tests
* Extract Source stubs into a reusable module
* Implement a helpers module to set up the global interface
